server/requirements.txt
boto3==1.13.6
botocore==1.16.6
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
docutils==0.15.2
future==0.18.2
idna==2.9
jmespath==0.9.5
line-bot-sdk==1.16.0
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==12.7
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
tornado==6.0.4
urllib3==1.25.9
